move TestLogDBGetter to tag-constrainted test file
