+++ raftstore/store_extension_test.go
@@ -1,0 +1,31 @@
+// Copyright 2021 MatrixOrigin.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build matrixone_test
+// +build matrixone_test
+
+package raftstore
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLogDBGetter(t *testing.T) {
+	// we just check whether this can compile
+	var s interface{}
+	s = &store{}
+	getter := s.(LogDBGetter)
+	assert.Nil(t, getter.GetLogDB())
+}
